requirements.txt
faiss-cpu
sentence-transformers
numpy
openai
python-dotenv
fastapi
uvicorn
